requirements.txt
fastapi
uvicorn
python-multipart
pyJWT
sqlmodel
mysql-connector
pymysql
asyncpg
pydantic-settings